backend/requirements.txt
fastapi
uvicorn
selenium
python-dotenv
langdetect
nltk
webdriver_manager
spacy